%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{AromaAffymetrixWrappers}
\alias{fixSearchPath.AromaAffymetrixWrappers}
\alias{extractSnpCnvQSet}
\alias{extractSnpCnvQSet.SnpChipEffectSet}
\alias{extractSnpQSet}
\alias{extractSnpQSet.SnpChipEffectSet}
\alias{extractTheta.SnpCnvQSet}
\alias{extractTheta.SnpQSet}
\alias{justSNPRMA}
\alias{justSNPRMA.AffymetrixCelSet}
\alias{justSNPRMA.character}
\alias{allocate.AromaCrlmmBinaryFile}
\alias{justCRLMMv2}
\alias{justCRLMMv2.SnpChipEffectSet}
\alias{byName.AromaCrlmmBinarySet}
\alias{CrlmmModel}
\alias{extractCalls}
\alias{extractCalls.AromaCrlmmBinaryFile}
\alias{extractCalls.AromaCrlmmBinarySet}
\alias{findUnitsTodo.AromaCrlmmBinaryFile}
\alias{findUnitsTodo.AromaCrlmmBinarySet}
\alias{findUnitsTodo.CrlmmModel}
\alias{fit.CrlmmModel}
\alias{getCallSet}
\alias{getCallSet.CrlmmModel}
\alias{getChipType.CrlmmModel}
\alias{getCrlmmPriors}
\alias{getCrlmmPriors.CrlmmModel}
\alias{getCrlmmSNPs}
\alias{getCrlmmSNPs.CrlmmModel}
\alias{getCrlmmSNPsOnChrX}
\alias{getCrlmmSNPsOnChrX.CrlmmModel}
\alias{getParameterSet.CrlmmModel}
\alias{getPlatformDesignDB}
\alias{getPlatformDesignDB.CrlmmModel}
\alias{getUnitsToFit.CrlmmModel}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
